use Proxy pattern to communicate with MSF over MessagePack

+++ msf/src/main/java/org/csploit/msf/impl/options/IntegerOption.java
@@ -5,22 +5,22 @@ import org.csploit.msf.impl.Option;
 /**
  * Integer option.
  */
-public class IntegerOption extends Option<Integer> {
-  public IntegerOption(String name, boolean required, String description, Integer defaultValue) {
+public class IntegerOption extends Option<Long> {
+  public IntegerOption(String name, boolean required, String description, Long defaultValue) {
     super(name, required, description, defaultValue);
   }
 
-  public IntegerOption(String name, String description, Integer defaultValue) {
+  public IntegerOption(String name, String description, Long defaultValue) {
     super(name, description, defaultValue);
   }
 
   @Override
-  protected Integer normalize(String input) {
+  protected Long normalize(String input) {
     try {
       if (input.startsWith("0x")) {
-        return Integer.parseInt(input.substring(2), 16);
+        return Long.parseLong(input.substring(2), 16);
       } else {
-        return Integer.parseInt(input);
+        return Long.parseLong(input);
       }
     } catch (NumberFormatException e) {
       return null;
